ограничил отказ взятия шмоток только рентерами и продавцами
