+++ src/cmd/give.cpp
@@ -17,8 +17,9 @@ void perform_give(CharData *ch, CharData *vict, ObjData *obj) {
 			false, ch, nullptr, nullptr, kToChar);
 		return;
 	}
-	if (NPC_FLAGGED(vict, ENpcFlag::kNoTakeItems) 
-			|| mob_index[GET_MOB_RNUM(vict)].func) {
+	if (vict->IsNpc() && (NPC_FLAGGED(vict, ENpcFlag::kNoTakeItems)
+			|| mob_index[GET_MOB_RNUM(vict)].func == shop_ext
+			|| mob_index[GET_MOB_RNUM(vict)].func == receptionist)) {
 		act("$N не нуждается в ваших подачках, своего барахла навалом.",
 			false, ch, nullptr, vict, kToChar);
 		return;
