+++ src/game_economics/ext_money.h
@@ -36,8 +36,6 @@ bool need_torc(CharData *ch);
 
 } // namespace Remort
 
-int torc(CharData *ch, void *me, int cmd, char *argument);
-
 #endif // EXT_MONEY_HPP_INCLUDED
 
 // vim: ts=4 sw=4 tw=0 noet syntax=cpp :
