+++ src/utils/utils.h
@@ -111,6 +111,7 @@ extern char KoiToWin2[];
 extern char AltToLat[];
 
 // public functions in utils.cpp
+std::string print_special(CharData *mob);
 CharData *find_char(long n);
 CharData *find_pc(long n);
 char *rustime(const struct tm *timeptr);
@@ -125,6 +126,9 @@ int strn_cmp(const std::string &arg1, const char *arg2, size_t n);
 int strn_cmp(const char *arg1, const std::string &arg2, size_t n);
 int strn_cmp(const std::string &arg1, const std::string &arg2, size_t n);
 int touch(const char *path);
+int exchange(CharData *ch, void *me, int cmd, char *argument);
+int horse_keeper(CharData *ch, void *me, int cmd, char *argument);
+int torc(CharData *ch, void *me, int cmd, char *argument);
 
 int number(int from, int to);
 int RollDices(int number, int size);
